Initial commit
Initial commit of Launchpad sequencer using python rtmidi bindings. An array of states is kept for each button in the main grid of the Launchpad, and steps can be toggled. A single element tracks through the grid at the BPM specified as the beginning of a sequencer. No MIDI note output has been established yet.