Add note muting
Pressing the scene trigger buttons next to the rows in a sequnce will mute the steps in that sequence. The steps will turn red and will be toggleable, but will not trigger notes when the sequencer reaches them.

The tick and update methods in lpview have been combined as they implement the same functionality.